Refactor #497 - rename core.kotlinx.coroutine to core.kotlinx.coroutines
