+++ core/androidx/compose/build.gradle.kts
@@ -11,7 +11,7 @@ dependencies {
     api(projects.core.kotlinx.datetime)
 
     implementation(projects.core.designsystem)
-    implementation(projects.core.kotlinx.coroutineAndroid)
+    implementation(projects.core.kotlinx.coroutinesAndroid)
 
     implementation(libs.androidx.compose.material3)
 }
