+++ core/androidx/compose/src/main/kotlin/com/nlab/reminder/core/androidx/compose/ui/ThrottleClick.kt
@@ -19,7 +19,7 @@ package com.nlab.reminder.core.androidx.compose.ui
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.*
-import com.nlab.reminder.core.kotlinx.coroutine.flow.throttleFirst
+import com.nlab.reminder.core.kotlinx.coroutines.flow.throttleFirst
 import kotlinx.coroutines.flow.MutableSharedFlow
 
 /**
