+++ core/androidx/lifecycle/build.gradle.kts
@@ -23,6 +23,6 @@ android {
 }
 
 dependencies {
-    api(projects.core.kotlinx.coroutineAndroid)
+    api(projects.core.kotlinx.coroutinesAndroid)
     api(libs.androidx.lifecycle.runtime.ktx)
 }
