+++ core/data-impl/src/main/kotlin/com/nlab/reminder/core/data/repository/impl/OfflineFirstLinkMetadataRepository.kt
@@ -24,9 +24,9 @@ import com.nlab.reminder.core.kotlin.collections.toSet
 import com.nlab.reminder.core.kotlin.concurrent.atomics.updateAndGet
 import com.nlab.reminder.core.kotlin.map
 import com.nlab.reminder.core.kotlin.onSuccess
-import com.nlab.reminder.core.kotlinx.coroutine.flow.channelFlow
-import com.nlab.reminder.core.kotlinx.coroutine.flow.combine
-import com.nlab.reminder.core.kotlinx.coroutine.flow.filter
+import com.nlab.reminder.core.kotlinx.coroutines.flow.channelFlow
+import com.nlab.reminder.core.kotlinx.coroutines.flow.combine
+import com.nlab.reminder.core.kotlinx.coroutines.flow.filter
 import com.nlab.reminder.core.local.database.dao.LinkMetadataDAO
 import com.nlab.reminder.core.network.datasource.LinkThumbnailDataSource
 import kotlinx.collections.immutable.persistentHashMapOf
