+++ core/kotlinx/coroutines-android/build.gradle.kts
@@ -3,10 +3,10 @@ plugins {
 }
 
 android {
-    namespace = "com.nlab.reminder.core.kotlinx.coroutine.android"
+    namespace = "com.nlab.reminder.core.kotlinx.coroutines.android"
 }
 
 dependencies {
-    api(projects.core.kotlinx.coroutine)
+    api(projects.core.kotlinx.coroutines)
     api(libs.kotlinx.coroutines.android)
 }
