+++ core/kotlinx/coroutines/src/main/java/com/nlab/reminder/core/kotlinx/coroutines/flow/Merge.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.kotlinx.coroutine.flow
+package com.nlab.reminder.core.kotlinx.coroutines.flow
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flatMapLatest as kotlinFlatMapLatest
