+++ core/network/build.gradle.kts
@@ -10,7 +10,7 @@ android {
 dependencies {
     api(projects.core.kotlin)
     api(projects.core.kotlinx.datetime)
-    implementation(projects.core.kotlinx.coroutine)
+    implementation(projects.core.kotlinx.coroutines)
 
     implementation(libs.google.play.services.time)
     implementation(libs.jsoup)
