+++ settings.gradle.kts
@@ -78,8 +78,8 @@ include(
     ":core:inject",
     ":core:kotlin",
     ":core:kotlin-test",
-    ":core:kotlinx:coroutine",
-    ":core:kotlinx:coroutine-android",
+    ":core:kotlinx:coroutines",
+    ":core:kotlinx:coroutines-android",
     ":core:kotlinx:datetime",
     ":core:local",
     ":core:network",
